# openapi.yaml
openapi: 3.0.3
info:
  title: Ahha Voucher Engine API
  description: |
    Multi-tenant SaaS backend for voucher distribution and redemption.

    ## Authentication

    This API uses two layers of authentication:
    1. **API Key** (`x-api-key` header) - Required for all requests
    2. **JWT Bearer Token** (`Authorization` header) - Required for admin endpoints

    ## Multi-Tenancy

    All requests must include the tenant subdomain in the Host header (e.g., `client1.ahha-be.local`).
    The subdomain determines which tenant database to use.
  version: 1.0.0
  contact:
    name: Ahha Team
servers:
  - url: http://{subdomain}.ahha-be.local
    variables:
      subdomain:
        default: client1
        description: Tenant subdomain
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - ApiKeyAuth: []
    BearerAuth: []
...
paths:
...
  /loyalty-admin/vouchers/{id}:
    get:
      tags:
        - Vouchers
      summary: Get Voucher by ID
      description: Retrieve a specific voucher (requires read:vouchers permission)
      operationId: findOneVoucher
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Voucher found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
        '404':
          description: Voucher not found

    patch:
      tags:
        - Vouchers
      summary: Update Voucher
      description: Update a voucher (requires write:vouchers permission)
      operationId: updateVoucher
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoucherDto'
      responses:
        '200':
          description: Voucher updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'

    delete:
      tags:
        - Vouchers
      summary: Delete Voucher
      description: Delete a voucher (requires write:vouchers permission)
      operationId: removeVoucher
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Voucher deleted successfully

  # ============ VOUCHER CATEGORIES ============
  /loyalty-admin/voucher-categories:
    get:
      tags:
        - Voucher Categories
      summary: Get All Voucher Categories
      description: Retrieve all voucher categories with pagination (requires read:voucher-categories permission)
      operationId: findAllVoucherCategories
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: Paginated list of voucher categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVoucherCategoryResponse'

    post:
      tags:
        - Voucher Categories
      summary: Create Voucher Category
      description: Create a new voucher category (requires create:voucher-categories permission)
      operationId: createVoucherCategory
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoucherCategoryDto'
      responses:
        '201':
          description: Voucher category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoucherCategory'
  /loyalty-admin/voucher-categories/{id}:
    get:
      tags:
        - Voucher Categories
      summary: Get Voucher Category by ID
      description: Retrieve a specific voucher category (requires read:voucher-categories permission)
      operationId: findOneVoucherCategory
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Voucher category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoucherCategory'

    put:
      tags:
        - Voucher Categories
      summary: Update Voucher Category
      description: Update a voucher category (requires update:voucher-categories permission)
      operationId: updateVoucherCategory
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoucherCategoryDto'
      responses:
        '200':
          description: Voucher category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoucherCategory'

    delete:
      tags:
        - Voucher Categories
      summary: Delete Voucher Category
      description: Delete a voucher category (requires delete:voucher-categories permission)
      operationId: removeVoucherCategory
      security:
        - BearerAuth: []
          ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Voucher category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ============ QUESTS ============
...
  /redistro/products/{id}:
    get:
      tags:
        - Products
      summary: Get Product by ID
      description: Retrieve a specific product
      operationId: findOneProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    patch:
      tags:
        - Products
      summary: Update Product
      description: Update a product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags:
        - Products
      summary: Delete Product
      description: Delete a product
      operationId: removeProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted successfully

  # ============ RETAILERS ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for the tenant
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: string
          description: Response code
        message:
          type: string
          description: Response message
        data:
          type: object
          description: Response data
    PaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: integer
                total:
                  type: integer
                size:
                  type: integer
            data:
              type: array
              items:
                type: object

    # ============ AUTH SCHEMAS ============
...
    VoucherBindingType:
      type: string
      enum:
        - role
        - product_type
        - product_sku
        - product_vendor
      description: Type of entity the voucher is bound to
    VoucherValidityType:
      type: string
      enum:
        - daily
        - birthday
        - weekly
        - monthly
        - one_time
      description: Type of voucher validity period
    VoucherCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          description: Unique identifier for the category
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
        image:
          type: string
          description: Category image URL
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateVoucherCategoryDto:
      type: object
      required:
        - slug
        - name
        - description
        - image
      properties:
        slug:
          type: string
          description: Unique slug for the category
          example: food-and-beverage
          minLength: 1
          maxLength: 100
        name:
          type: string
          description: Category name
          example: Food & Beverage
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Category description
          example: Food and beverage vouchers for restaurants, cafes, and bars
          minLength: 1
          maxLength: 500
        image:
          type: string
          description: Image URL for the category
          example: https://example.com/images/food-beverage.png
          minLength: 1
          maxLength: 255
    UpdateVoucherCategoryDto:
      allOf:
        - $ref: '#/components/schemas/CreateVoucherCategoryDto'
    VoucherValidity:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: '#/components/schemas/VoucherValidityType'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        start_time:
          type: string
          example: '00:00:00'
        end_time:
          type: string
          example: '23:59:59'
    CreateVoucherValidityDto:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/VoucherValidityType'
        start_date:
          type: string
          format: date-time
          description: Start date in ISO 8601 format
        end_date:
          type: string
          format: date-time
          description: End date in ISO 8601 format
        start_time:
          type: string
          description: Start time (HH:MM:SS)
          example: '09:00:00'
        end_time:
          type: string
          description: End time (HH:MM:SS)
          example: '18:00:00'
    VoucherBinding:
      type: object
      properties:
        id:
          type: integer
        bind_type:
          type: string
          description: Type of binding
        bind_value:
          type: string
          description: Value of the binding
    CreateVoucherBindingDto:
      type: object
      required:
        - bind_type
        - bind_value
      properties:
        bind_type:
          $ref: '#/components/schemas/VoucherBindingType'
        bind_value:
          type: string
          description: Value to bind to (e.g., role name, product SKU)
    Voucher:
      type: object
      properties:
        code:
          type: string
          description: Unique voucher code (primary key)
        description:
          type: string
          description: Voucher description
        quota:
          type: integer
          description: Maximum number of times this voucher can be claimed
        image:
          type: string
          description: Voucher image URL
        categories:
          type: array
          items:
            $ref: '#/components/schemas/VoucherCategory'
        allow_combine_categories:
          type: array
          items:
            $ref: '#/components/schemas/VoucherCategory'
        validities:
          type: array
          items:
            $ref: '#/components/schemas/VoucherValidity'
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/VoucherBinding'
        target_users:
          type: array
          items:
            type: string
            format: uuid
          description: UUIDs of users allowed to use this voucher
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateVoucherDto:
      type: object
      required:
        - code
        - quota
        - categories
        - allow_combine_categories
        - validities
        - bindings
      properties:
        code:
          type: string
          description: Unique voucher code
        description:
          type: string
          description: Voucher description
        quota:
          type: integer
          description: Maximum claim count
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CreateVoucherCategoryDto'
          description: Categories this voucher belongs to
        allow_combine_categories:
          type: array
          items:
            $ref: '#/components/schemas/CreateVoucherCategoryDto'
          description: Categories this voucher can be combined with
        validities:
          type: array
          items:
            $ref: '#/components/schemas/CreateVoucherValidityDto'
          description: Validity periods for this voucher
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/CreateVoucherBindingDto'
          description: Product/brand/store bindings
        target_users:
          type: array
          items:
            type: string
            format: uuid
          description: UUIDs of target users (optional)
    UpdateVoucherDto:
      allOf:
        - $ref: '#/components/schemas/CreateVoucherDto'
...
    PaginatedVoucherCategoryResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/VoucherCategory'

    # ============ CONSUMER VOUCHER SCHEMAS ============
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          description: Product price
        sku:
          type: string
          description: Product SKU
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProductDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          description: Product price
        sku:
          type: string
          description: Product SKU
    UpdateProductDto:
      allOf:
        - $ref: '#/components/schemas/CreateProductDto'

    # ============ RETAILER SCHEMAS ============
...